feat: completly functionality
